services/notification_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
pika==1.3.1
django-cors-headers==4.3.1
drf-yasg==1.21.5
pyyaml>=6.0
uritemplate>=4.1.1 